Feature Visualization/requirements.txt
Flask==3.0.3
joblib==1.4.2
matplotlib==3.9.2
nibabel==5.3.2
nnunetv2==2.5.1
numpy==2.1.2
pandas==2.2.3
pyradiomics==3.0.1
scikit_learn==1.5.2
shap==0.46.0
SimpleITK==2.4.0
torch==2.5.1
